requirements.txt
python-sonarqube-api
requests
python-dotenv